Create Decommenter
This program is meant to mimic some of the functionality of the preprocessor, specifically in removing comments and leaving behind a single space if one was not their previously. The code also alerts the user if their is an unterminated comment at the end. Currently the code does not handle nested quotations, does not handle runaway comments that are closed before the end of the program, and it does not alert the user when a runaway comment begins.